learn::evol
Module phenotype

Phenotype Trait

The Phenotype trait defines the interface for types that represent individuals in an evolutionary algorithm. It provides methods for crossover and mutation.

Example

use learn::evol::phenotype::Phenotype;
use learn::evol::rng::RandomNumberGenerator;

#[derive(Clone, Debug)]
struct MyPhenotype {
    // ... fields and methods for your specific phenotype ...
}

impl Phenotype for MyPhenotype {
    fn crossover(&mut self, other: &Self) {
        // Implementation of crossover for MyPhenotype
        // ...
    }

    fn mutate(&mut self, rng: &mut RandomNumberGenerator) {
        // Implementation of mutation for MyPhenotype
        // ...
    }
}

Trait

Phenotype

The Phenotype trait defines methods for crossover and mutation.

Methods

crossover(&self, other: &Self)

Performs crossover with another individual of the same type.

mutate(&mut self, rng: &mut RandomNumberGenerator)

Performs mutation on the individual using the provided random number generator.

Implementing the Trait

To use the Phenotype trait, implement it for your custom phenotype type.

use learn::evol::phenotype::Phenotype;
use learn::evol::rng::RandomNumberGenerator;

#[derive(Clone, Debug)]
struct MyPhenotype {
    // ... fields and methods for your specific phenotype ...
}

impl Phenotype for MyPhenotype {
    fn crossover(&mut self, other: &Self) {
        // Implementation of crossover for MyPhenotype
        // ...
    }

    fn mutate(&mut self, rng: &mut RandomNumberGenerator) {
        // Implementation of mutation for MyPhenotype
        // ...
    }
}

Traits

Phenotype