learn::evol::strategy
Trait BreedStrategy

pub trait BreedStrategy<Pheno: Phenotype>
where
    Self: Debug + Clone,
{
    // Required method
    fn breed(
        &self,
        parents: &[Pheno],
        evol_options: &EvolutionOptions,
        rng: &mut RandomNumberGenerator,
    ) -> Result<Vec<Pheno>, Error>;
}

BreedStrategy

The BreedStrategy trait defines the interface for strategies responsible for breeding new individuals (phenotypes) based on a set of parent individuals and evolution options.

Required Methods

fn breed( &self, parents: &[Pheno], evol_options: &EvolutionOptions, rng: &mut RandomNumberGenerator, ) -> Result<Vec<Pheno>, Error>

Breeds new individuals based on a set of parent individuals and evolution options. The breed method is responsible for generating a new population of individuals based on a set of parent individuals, evolution options, and a random number generator.

Parameters

parents: A slice containing the parent individuals.

evol_options: A reference to the evolution options specifying algorithm parameters.

rng: A mutable reference to the random number generator used for generating random values during breeding.

Returns

A vector containing the newly bred individuals.

Object Safety

This trait is not object safe.