learn::neural::layer::layer_trait
Trait Layer

pub trait Layer: Debug + DynClone {

    // Required methods
    fn forward(&mut self, input: &[f64]) -> Vec<f64>;
    fn forward_batch(&mut self, input: &[f64]) -> Vec<f64>;
    fn backward(&mut self, grad_output: &[f64]) -> Vec<f64>;
    fn backward_batch(&mut self, grad_output: &[f64]) -> Vec<f64>;
    fn update_weights(&mut self, learning_rate: f64);
    fn input_size(&self) -> usize;
    fn output_size(&self) -> usize;
    fn save(&self, path: &str) -> Result<(), Box<dyn Error>>;
    fn read(&mut self, path: &str) -> Result<(), Box<dyn Error>>;
    fn resize(&mut self, input_size: usize, output_size: usize);
    fn assign_weights(&mut self, other: &dyn Layer);
    fn get_weights(&self) -> Matrix<f64>;
    fn get_biases(&self) -> Vec<f64>;
    fn adjust_adam(
        &mut self,
        t: usize,
        learning_rate: f64,
        beta1: f64,
        beta2: f64,
        epsilon: f64,
    );

}

Provides methods for the forward pass, backward pass, weight updates, and layer size information.

Required Methods

fn forward(&mut self, input: &[f64]) -> Vec<f64>

Performs the forward pass of the layer, computing the output based on the input vector.

Arguments

input - A reference to a vector of f64 values representing the input data.

Returns

A vector of f64 values representing the output of the layer.

fn forward_batch(&mut self, input: &[f64]) -> Vec<f64>

Performs the forward pass of the layer for inputs doing batch caching.

fn backward(&mut self, grad_output: &[f64]) -> Vec<f64>

Performs the backward pass of the layer, computing the gradient based on the output gradient.

Arguments

grad_output - A reference to a vector of f64 values representing the gradient of the loss with respect to the output of this layer.

Returns

A vector of f64 values representing the gradient of the loss with respect to the input.

fn backward_batch(&mut self, grad_output: &[f64]) -> Vec<f64>

Performs the backward pass of the layer for inputs doing batch caching.

fn update_weights(&mut self, learning_rate: f64)

Updates the weights of the layer based on the specified learning rate.

Arguments

learning_rate - A f64 value representing the learning rate for weight updates.

fn input_size(&self) -> usize

Returns the input size of the layer.

Returns

A usize value representing the number of input neurons.

fn output_size(&self) -> usize

Returns the output size of the layer.

Returns

A usize value representing the number of output neurons.

fn save(&self, path: &str) -> Result<(), Box<dyn Error>>

Saves the layer to a file at the specified path.

fn read(&mut self, path: &str) -> Result<(), Box<dyn Error>>

Reads the layer from a file at the specified path.

fn resize(&mut self, input_size: usize, output_size: usize)

Resizes the layer to the input dimensions.

fn assign_weights(&mut self, other: &dyn Layer)

Assigns the weight of the input other layer

fn get_weights(&self) -> Matrix<f64>

Returns the weights of the layer.

fn get_biases(&self) -> Vec<f64>

Returns the biases of the layer.

fn adjust_adam( &mut self, t: usize, learning_rate: f64, beta1: f64, beta2: f64, epsilon: f64, )

Adjusts the weights according to the Adam optimizer.